pyproject.toml
[project]
name = "index"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.3.0",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "sentence-transformers>=4.1.0",
    "uvicorn>=0.34.3",
]
